refactor: 바뀐 컴포넌트 경로에 맞게 import문 수정 (#230)
