+++ src/pages/HomePage/Box/BoxTodayTodo/StatusDefaultBoxTodayTodo/StatusDefaultBoxTodayTodo.tsx
@@ -1,7 +1,8 @@
-import HomeLargeBtn from '@/components/atoms/HomeLargeBtn';
+import HomeLargeBtn from '@/shared/components/Button/ButtonHomeLarge/ButtonHomeLarge';
+
+import { HomeLargeBtnVariant } from '@/shared/types/global';
 
 import { LARGE_BTN_TEXT } from '@/shared/constants/btnText';
-import { HomeLargeBtnVariant } from '@/shared/types/global';
 
 interface StatusDefaultBoxTodayTodoProps {
 	hasTodos: boolean;
@@ -11,8 +12,8 @@ interface StatusDefaultBoxTodayTodoProps {
 const StatusDefaultBoxTodayTodo = ({ hasTodos, onEnableAddStatus }: StatusDefaultBoxTodayTodoProps) => {
 	return (
 		<div className="mt-[21.4rem] flex w-full flex-col justify-center text-white">
-			<p className="subhead-med-18 text-center text-gray-03">아직 오늘 할 일이 없어요</p>
-			<p className="subhead-bold-22 mb-[2.2rem] mt-[1.2rem] text-center text-gray-05">
+			<p className="text-center text-gray-03 subhead-med-18">아직 오늘 할 일이 없어요</p>
+			<p className="mb-[2.2rem] mt-[1.2rem] text-center text-gray-05 subhead-bold-22">
 				할 일을 추가해
 				<br />
 				타이머를 시작해보세요.
