+++ src/pages/HomePage/StatusDefaultHome/StatusDefaultHome.tsx
@@ -1,10 +1,10 @@
+import HomeLargeBtn from '@/shared/components/Button/ButtonHomeLarge/ButtonHomeLarge';
+
 import { HomeLargeBtnVariant } from '@/shared/types/global';
 
 import { LARGE_BTN_TEXT } from '@/shared/constants/btnText';
 
 import HomeDefaultIcon from '@/shared/assets/svgs/home_default_icon.svg?react';
-
-import HomeLargeBtn from '@/components/atoms/HomeLargeBtn';
 
 interface StatusDefaultHomeProps {
 	onClick?: () => void;
