+++ src/pages/OnboardingPage/components/Step/StepStart/StepStart.tsx
@@ -1,8 +1,8 @@
+import HomeLargeBtn from '@/shared/components/Button/ButtonHomeLarge/ButtonHomeLarge';
+
 import { HomeLargeBtnVariant } from '@/shared/types/global';
 
 import { OnboardingIcon } from '@/shared/assets/home/svgs';
-
-import HomeLargeBtn from '@/components/atoms/HomeLargeBtn';
 
 interface StepStartProps {
 	setStep: (step: string) => void;
@@ -11,8 +11,8 @@ interface StepStartProps {
 const StepStart = ({ setStep }: StepStartProps) => {
 	return (
 		<main className="flex w-full flex-col items-center justify-center">
-			<h1 className="title-bold-36 mb-[2rem] text-center text-white">집중을 도와줄 모립세트를 만들어볼까요?</h1>
-			<p className="body-reg-24 mb-[8.3rem] text-center text-gray-04">
+			<h1 className="mb-[2rem] text-center text-white title-bold-36">집중을 도와줄 모립세트를 만들어볼까요?</h1>
+			<p className="mb-[8.3rem] text-center text-gray-04 body-reg-24">
 				모립세트란 작업에 꼭 필요한 일만 할 수 있도록
 				<br />
 				설정하는 작업 시 사용 할 허용 서비스들의 모음을 말해요.
@@ -23,7 +23,7 @@ const StepStart = ({ setStep }: StepStartProps) => {
 				<span className="min-w-[15.6rem]">시작하기</span>
 			</HomeLargeBtn>
 
-			<button className="subhead-reg-22 text-gray-04">건너뛰기</button>
+			<button className="text-gray-04 subhead-reg-22">건너뛰기</button>
 		</main>
 	);
 };
