+++ src/pages/TimerPage/components/Carousel.tsx
@@ -1,8 +1,9 @@
+import ArrowSVGBtn from '@/shared/components/Button/ButtonArrowSVG/ButtonArrowSVG';
+
 import useFriendInfoCarousel from '@/shared/hooks/useFriendInfoCarousel';
 
 import { Direction } from '@/shared/types/global';
 
-import ArrowSVGBtn from '@/components/atoms/ArrowSVGBtn';
 import FriendInfo from '@/pages/TimerPage/components/CarouselContainer';
 
 const Carousel = () => {
